// src/generated/models/Cr9b5_pt_referencesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9b5_pt_referencescr9b5_referencetype = {
  233100000: 'Property',
  233100001: 'Supplier',
  233100002: 'Client',
  233100003: 'IncomingInvoice',
  233100004: 'OutgoingInvoice'
} as const;
export type Cr9b5_pt_referencescr9b5_referencetype = keyof typeof Cr9b5_pt_referencescr9b5_referencetype;
export const Cr9b5_pt_referencesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9b5_pt_referencesstatecode = keyof typeof Cr9b5_pt_referencesstatecode;
export const Cr9b5_pt_referencesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9b5_pt_referencesstatuscode = keyof typeof Cr9b5_pt_referencesstatuscode;

export interface Cr9b5_pt_referencesBase {
  cr9b5_pt_referenceid: string;
  cr9b5_referencetype?: Cr9b5_pt_referencescr9b5_referencetype;
  cr9b5_sortorder?: number;
  cr9b5_value: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9b5_pt_referencesstatecode;
  statuscode?: Cr9b5_pt_referencesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9b5_pt_references extends Cr9b5_pt_referencesBase {
  cr9b5_referencetypename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}